+++ src/qumin/cli.py
@@ -23,8 +23,12 @@ def qumin_command(cfg):
         not_overab = cfg.pats.overabundant is False
         not_defect = cfg.pats.defective is False
         for_m = cfg.action == "macroclasses"
-        assert not_overab or not for_m, "For this calculation, overabundant must be False"
-        assert not_defect or not for_m, "For this calculation, defective must be False"
+        for_H = cfg.action == "H"
+        assert not_overab or not for_m, "For this calculation, pats.overabundant must be False"
+        assert not_defect or not for_m, "For this calculation, pats.defective must be False"
+        assert not_overab or not for_H or cfg.entropy.use_extra, \
+            "For this calculation, pats.overabundant must be False " \
+            "or entropy.use_extra must be True."
         patterns_file = pat_command(cfg, md)
         cfg.patterns = patterns_file