+++ src/qumin/find_patterns.py
@@ -52,9 +52,6 @@ def pat_command(cfg, md):
                            gap_prop=cfg.pats.gap_proportion,
                            cpus=cfg.cpus or min(1, cpu_count() - 2))
 
-    # Concatenate the patterns as a dict. Cell names are turned into columns.
-    # patterns_df = pd.concat([df for df in patterns_dfs.values()]).reset_index(drop=True)
-
     if merge_cols and not cfg.pats.merged:  # Re-build duplicate columns
         paradigms.unmerge_columns()
         patterns.unmerge_columns(paradigms)
@@ -67,12 +64,13 @@ def pat_command(cfg, md):
                     "-- This means something went wrong."
                     "Please report this as a bug !")
 
-    microclasses = find_microclasses(paradigms, patterns)
-    filename = md.register_file("microclasses.txt",
-                                {'computation': cfg.pats.kind, 'content': 'microclasses'})
-    log.info("Found %s microclasses. Printing microclasses to %s", len(microclasses), filename)
-    with open(filename, "w", encoding="utf-8") as flow:
-        for m in sorted(microclasses, key=lambda m: len(microclasses[m])):
-            flow.write("\n\n{} ({}) \n\t".format(m, len(microclasses[m])) + ", ".join(microclasses[m]))
+    # microclasses = find_microclasses(paradigms, patterns)
+    # filename = md.register_file("microclasses.txt",
+    #                             {'computation': cfg.pats.kind, 'content': 'microclasses'})
+    # log.info("Found %s microclasses. Printing microclasses to %s", len(microclasses), filename)
+    # with open(filename, "w", encoding="utf-8") as flow:
+    #     for m in sorted(microclasses, key=lambda m: len(microclasses[m])):
+    #         flow.write("\n\n{} ({}) \n\t".format(m, len(microclasses[m]))
+    #                    + ", ".join(microclasses[m]))
 
     return patterns.export(md, kind, optim_mem=cfg.pats.optim_mem)