+++ src/qumin/representations/frequencies.py
@@ -168,6 +168,7 @@ class Frequencies(object):
         # We save the resulting table
         table.sort_index(inplace=True)
         table.index.name = name[:-1]
+        table.index = table.index.astype('str')
         setattr(self, name, table[cols])
 
     def get_absolute_freq(self, mean=False, group_on=False, skipna=False, **kwargs):
@@ -262,6 +263,8 @@ class Frequencies(object):
             array([0.25, 0.25, 0.25, 0.25])
             >>> f.get_relative_freq(filters={'lexeme':'p'}, group_on=["lexeme", "cell"])['result'].values
             array([0.05882353, 0.94117647, 1.        , 1.        ])
+            >>> f.get_relative_freq(filters={'lexeme':'s', 'cell': 'first'}, group_on=["lexeme", "cell"]).result.values
+            array([0.33333333, 0.33333333, 0.33333333])
 
         Arguments:
             group_on (List[str]): column on which relative frequencies should be computed
@@ -289,12 +292,16 @@ class Frequencies(object):
         sublist.result = sublist.result.astype('float64')
 
         # 2. If there are any NaN values, we give a uniform frequency to the group
-        sublist['notna'] = True
-        sublist.loc[sublist.value.isna(), 'notna'] = False
-        nanval = (sublist.result != 1) & ~sublist.groupby(groups).notna.transform('all')
-        sublist.loc[nanval, 'result'] = 1/sublist.loc[nanval, 'result']
+        any_nan = sublist.groupby(groups).value.transform(lambda x: x.isna().any())
 
-        # 3. If all values are filled and if the group is bigger than one, we sum the frequencies
+        # 3. If a whole group contains zeros, we give a uniform frequency to the group
+        all_zero = sublist.groupby(groups).value.transform(lambda x: (x == 0).all())
+
+        # Apply 2 and 3
+        selector = (sublist.result != 1) & (any_nan | all_zero)
+        sublist.loc[selector, 'result'] = 1/sublist.loc[selector, 'result']
+
+        # 4. If all values are filled and if the group is bigger than one, we sum the frequencies
         selector = sublist.result > 1
 
         if group_on is False: